+++ src/parser/inline.test.ts
@@ -11,14 +11,11 @@ describe('Unit: parser/inline', () => {
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('++<a>`b`&c;&amp; ++')), [['<ins>&lt;a&gt;<code data-src="`b`">b</code>&amp;c;&amp; </ins>'], '']);
-      assert.deepStrictEqual(inspect(parser('~~<ruby>`a`</ruby>~~')), [['<del><ruby><code data-src="`a`">a</code></ruby></del>'], '']);
-      assert.deepStrictEqual(inspect(parser('<ruby>~~<mark>a</mark>~~</ruby>')), [['<ruby><del><mark>a</mark></del></ruby>'], '']);
-    });
-
-    it('flip', () => {
-      assert.deepStrictEqual(inspect(parser('**~~*<a>*~~**')), [['<strong><del><em>&lt;a&gt;</em></del></strong>'], '']);
-      assert.deepStrictEqual(inspect(parser('*~~**<a>**~~*')), [['<em><del><strong>&lt;a&gt;</strong></del></em>'], '']);
+      assert.deepStrictEqual(inspect(parser('**[*a*]**')), [['<strong>[<em>a</em>]</strong>'], '']);
+      assert.deepStrictEqual(inspect(parser('*(**a**)*')), [['<em>(<strong>a</strong>)</em>'], '']);
+      assert.deepStrictEqual(inspect(parser('*<a>`b`&c;&amp; *')), [['<em>&lt;a&gt;<code data-src="`b`">b</code>&amp;c;&amp; </em>'], '']);
+      assert.deepStrictEqual(inspect(parser('*<ruby>`a`</ruby>*')), [['<em><ruby><code data-src="`a`">a</code></ruby></em>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby>*<mark>a</mark>*</ruby>')), [['<ruby><em><mark>a</mark></em></ruby>'], '']);
     });
 
     it('link', () => {