+++ src/parser/inline.ts
@@ -2,12 +2,8 @@
 import { combine } from '../combinator/combine';
 import { annotation } from './inline/annotation';
 import { parenthesis } from './inline/parenthesis';
-import { insertion } from './inline/insertion';
-import { deletion } from './inline/deletion';
 import { strong } from './inline/strong';
 import { emphasis } from './inline/emphasis';
-import { superscript } from './inline/superscript';
-import { subscript } from './inline/subscript';
 import { code } from './inline/code';
 import { math } from './inline/math';
 import { media } from './inline/media';
@@ -22,12 +18,8 @@ import { text } from './source/text';
 export import InlineParser = MarkdownParser.InlineParser;
 export import AnnotationParser = InlineParser.AnnotationParser;
 export import ParenthesisParser = InlineParser.ParenthesisParser;
-export import InsertionParser = InlineParser.InsertionParser;
-export import DeletionParser = InlineParser.DeletionParser;
 export import StrongParser = InlineParser.StrongParser;
 export import EmphasisParser = InlineParser.EmphasisParser;
-export import SuperScriptParser = InlineParser.SuperScriptParser;
-export import SubScriptParser = InlineParser.SubScriptParser;
 export import CodeParser = InlineParser.CodeParser;
 export import MathInlineParser = InlineParser.MathInlineParser;
 export import MediaParser = InlineParser.MediaParser;
@@ -40,12 +32,8 @@ export import AutolinkParser = InlineParser.AutolinkParser;
 export const inline: InlineParser = combine<[
   AnnotationParser,
   ParenthesisParser,
-  InsertionParser,
-  DeletionParser,
   StrongParser,
   EmphasisParser,
-  SuperScriptParser,
-  SubScriptParser,
   CodeParser,
   MathInlineParser,
   MediaParser,
@@ -58,12 +46,8 @@ export const inline: InlineParser = combine<[
 ], HTMLElement | Text>([
   annotation,
   parenthesis,
-  insertion,
-  deletion,
   strong,
   emphasis,
-  superscript,
-  subscript,
   code,
   math,
   media,