+++ src/parser/inline/html.ts
@@ -7,10 +7,10 @@ import { squash } from '../squash';
 type SubParsers = [InlineParser];
 
 const syntax = /^(<([a-z]+)>)/;
-const inlinetags = Object.freeze('small|q|cite|mark|ruby|rt|rp|bdi|bdo|wbr'.split('|'));
+const inlinetags = Object.freeze('ins|del|sup|sub|small|q|cite|mark|ruby|rt|rp|bdi|bdo|wbr'.split('|'));
 assert(inlinetags.every(t => /[a-z]+/.test(t)));
 assert(inlinetags.every(t => ['script', 'style', 'link', 'a', 'img'].indexOf(t) === -1));
-assert(inlinetags.every(t => ['ins', 'del', 'strong', 'em', 'sup', 'sub', 'code', 's', 'u'].indexOf(t) === -1));
+assert(inlinetags.every(t => ['strong', 'em', 'code', 's', 'u'].indexOf(t) === -1));
 
 export const html: HTMLParser = function (source: string): Result<HTMLElement, SubParsers> {
   if (!source.startsWith('<')) return;